feat (theme): watch conf-img/tpl folder and never forget to generate sizes and locations json
