--- assets/conf-img/image-locations.json
@@ -1,28 +1,0 @@
-[
-  {
-    "square": [
-      {
-        "srcsets": [
-          {
-            "srcset": "",
-            "size": "img-100-100"
-          },
-          {
-            "srcset": "2x",
-            "size": "img-200-200"
-          },
-          {
-            "srcset": "",
-            "size": "img-200-200"
-          },
-          {
-            "srcset": "2x",
-            "size": "img-400-400"
-          }
-        ],
-        "default_img": "default-400-400.jpg",
-        "img_base": "img-400-400"
-      }
-    ]
-  }
-]
