--- assets/conf-img/image-sizes.json
@@ -1,19 +1,0 @@
-[
-  {
-    "img-100-100": {
-      "width": "100",
-      "height": "100",
-      "crop": true
-    },
-    "img-200-200": {
-      "width": "200",
-      "height": "200",
-      "crop": true
-    },
-    "img-400-400": {
-      "width": "400",
-      "height": "400",
-      "crop": true
-    }
-  }
-]
