+++ src/js/post-build.js
@@ -5,8 +5,5 @@ function requireAll(r) {
 // SVG
 requireAll(require.context('../img/icons', true, /\.svg$/))
 
-// DEFAULT
-requireAll(require.context('../img/default', true, /\.(png|jpe?g)$/))
-
 // STATIC
 requireAll(require.context('../img/static', true, /\.(png|jpe?g|gif)$/))
